Added pageobject for course overview and restructured a bit
