+++ src/test/cypress/support/pageobjects/course/CourseManagementPage.ts
